Move multiplayer controls widget classes each to a separate file
